+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 # Release Notes for Blogify
 
 ### Unreleased - 2021-11-24
+### Fixed 
+- Fixed author query
+
 ### Added
 - Added optional post views field to allow sorting by popularity
 ### Changed
 - Refactored code
+- Seeding: Randomize author
 
 ## 1.0.6 - 2021-11-23
 ### Improved